requirements.txt
flask==2.0.1
flask-socketio==5.3.2
mysql-connector-python==8.0.26
eventlet==0.33.1
redis==4.3.4
python-dotenv==0.19.0